+++ test/jdk/java/io/IO/script.exp
